requirements.txt
flake8==6.0.0
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytesseract==0.3.10
